Move light_all_transactions to light_fetch
