+++ rpc/src/v1/helpers/mod.rs
@@ -36,10 +36,6 @@ mod subscribers;
 mod subscription_manager;
 mod work;
 
-use std::sync::Arc;
-
-use transaction::PendingTransaction;
-
 pub use self::dispatch::{Dispatcher, FullDispatcher, LightDispatcher};
 pub use self::network_settings::NetworkSettings;
 pub use self::poll_manager::PollManager;
@@ -59,13 +55,4 @@ pub use self::work::submit_work_detail;
 
 pub fn to_url(address: &Option<::Host>) -> Option<String> {
 	address.as_ref().map(|host| (**host).to_owned())
-}
-
-pub fn light_all_transactions(dispatch: &Arc<LightDispatcher>) -> impl Iterator<Item=PendingTransaction> {
-	let txq = dispatch.transaction_queue.read();
-	let chain_info = dispatch.client.chain_info();
-
-	let current = txq.ready_transactions(chain_info.best_block_number, chain_info.best_block_timestamp);
-	let future = txq.future_transactions(chain_info.best_block_number, chain_info.best_block_timestamp);
-	current.into_iter().chain(future.into_iter())
 }
